+++ packages/moonstone/ProgressBar/ProgressBar.js
@@ -8,11 +8,14 @@
  * @exports ProgressBar
  * @exports ProgressBarBase
  * @exports ProgressBarDecorator
+ * @exports ProgressBarTooltip
  */
 
 import kind from '@enact/core/kind';
+import ComponentOverride from '@enact/ui/ComponentOverride';
 import UiProgressBar from '@enact/ui/ProgressBar';
 import Pure from '@enact/ui/internal/Pure';
+import Slottable from '@enact/ui/Slottable';
 import PropTypes from 'prop-types';
 import compose from 'ramda/src/compose';
 import React from 'react';
@@ -69,45 +72,38 @@ const ProgressBarBase = kind({
 		progress: PropTypes.number,
 
 		/**
-		 * Enables the built-in tooltip, whose behavior can be modified by the other tooltip
-		 * properties.
+		 * Enables the built-in tooltip
 		 *
-		 * @type {Boolean}
+		 * To customize the tooltip, pass either a custom Tooltip component or an instance of
+		 * [ProgressBarTooltip]{@link moonstone/ProgressBar.ProgressBarTooltip} with additional
+		 * props configured.
+		 *
+		 * ```
+		 * <ProgressBar
+		 *   tooltip={
+		 *     <ProgressBarTooltip side="after" />
+		 *   }
+		 * />
+		 * ```
+		 *
+		 * The tooltip may also be passed as a child via the `"tooltip"` slot. See
+		 * [Slottable]{@link ui/Slottable} for more information on how slots can be used.
+		 *
+		 * ```
+		 * <ProgressBar>
+		 *   <ProgressBarTooltip side="after" />
+		 * </ProgressBar>
+		 * ```
+		 *
+		 * @type {Boolean|Element|Function}
 		 * @public
 		 */
-		tooltip: PropTypes.bool,
-
-		/**
-		 * Setting to `true` overrides the natural tooltip position
-		 * for `vertical` progress bar to the left. This may be useful
-		 * if you have a static layout that have the progress bar at
-		 * the right edge of the container.
-		 *
-		 * @type {Boolean}
-		 * @public
-		 */
-		tooltipForceSide: PropTypes.bool,
-
-		/**
-		 * Specify where the tooltip should appear in relation to the ProgressBar. Options are
-		 * `'before'` and `'after'`. `before` renders above a `horizontal` slider and to the
-		 * left of a `vertical` ProgressBar. `after` renders below a `horizontal` slider and to the
-		 * right of a `vertical` ProgressBar. In the `vertical` case, the rendering position is
-		 * automatically reversed when rendering in an RTL locale. This can be overridden by
-		 * using the [tooltipForceSide]{@link moonstone/ProgressBar.ProgressBar.tooltipForceSide}
-		 * prop.
-		 *
-		 * @type {String}
-		 * @default 'before'
-		 * @public
-		 */
-		tooltipSide: PropTypes.string
+		tooltip: PropTypes.oneOfType([PropTypes.bool, PropTypes.object, PropTypes.func])
 	},
 
 	defaultProps: {
-		tooltip: false,
-		tooltipSide: 'before',
-		tooltipForceSide: false
+		orientation: 'horizontal',
+		progress: 0
 	},
 
 	styles: {
@@ -116,40 +112,26 @@ const ProgressBarBase = kind({
 	},
 
 	computed: {
-		tooltipComponent: ({orientation, progress, tooltip, tooltipForceSide, tooltipSide}) => {
-			if (tooltip) {
-				const progressAfterMidpoint = progress > 0.5;
-				const progressPercentage = Math.min(parseInt(progress * 100), 100);
-				const percentageText = `${progressPercentage}%`;
-
-				return (
-					<ProgressBarTooltip
-						forceSide={tooltipForceSide}
-						knobAfterMidpoint={progressAfterMidpoint}
-						proportion={progress}
-						orientation={orientation}
-						side={tooltipSide}
-					>
-						{percentageText}
-					</ProgressBarTooltip>
-				);
-			} else {
-				return null;
-			}
-		}
+		tooltip: ({tooltip}) => tooltip === true ? ProgressBarTooltip : tooltip
 	},
 
-	render: ({css, tooltipComponent, ...rest}) => {
+	render: ({css, orientation, progress, tooltip, ...rest}) => {
 		delete rest.tooltip;
-		delete rest.tooltipSide;
-		delete rest.tooltipForceSide;
 
 		return (
 			<UiProgressBar
 				{...rest}
+				orientation={orientation}
+				progress={progress}
 				css={css}
 			>
-				{tooltipComponent}
+				<ComponentOverride
+					component={tooltip}
+					orientation={orientation}
+					percent
+					proportion={progress}
+					visible
+				/>
 			</UiProgressBar>
 		);
 	}
@@ -165,6 +147,7 @@ const ProgressBarBase = kind({
  */
 const ProgressBarDecorator = compose(
 	Pure,
+	Slottable({slots: ['tooltip']}),
 	Skinnable
 );